# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/phones": {
            "get": {
                "description": "Get a list of Phone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Get all Phone.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Phone"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Phone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Create New Phone.",
                "parameters": [
                    {
                        "description": "the body to create a new Phone",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.phoneInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger: 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Phone"
                        }
                    }
                }
            }
        },
        "/phones/bybrand/{id}": {
            "get": {
                "description": "Get a list of Phone By Brand id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Get all Phone By Brand id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Phone"
                            }
                        }
                    }
                }
            }
        },
...
        "/phones/search/{keyword}": {
            "get": {
                "description": "Get a list of Phone By Name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Get all Phone By Name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword",
                        "name": "keyword",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Phone"
                            }
                        }
                    }
                }
            }
        },
        "/phones/{id}": {
            "get": {
                "description": "Get detail of Phone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Get Detail Phone.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Phone",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Phone"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Phone by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Delete one Phone.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger: 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Phone by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phone"
                ],
                "summary": "Update Phone.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age rating category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.phoneInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger: 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Phone"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.phoneInput": {
            "type": "object",
            "properties": {
                "battery_charging": {
                    "type": "string"
                },
                "battery_type": {
                    "type": "string"
                },
                "body_dimensions": {
                    "type": "string"
                },
                "body_sim": {
                    "type": "string"
                },
                "body_weight": {
                    "type": "string"
                },
                "brandID": {
                    "type": "integer"
                },
                "comms_bluetooth": {
                    "type": "string"
                },
                "comms_gps": {
                    "type": "string"
                },
                "comms_infrared": {
                    "type": "string"
                },
                "comms_nfc": {
                    "type": "string"
                },
                "comms_radio": {
                    "type": "string"
                },
                "comms_usb": {
                    "type": "string"
                },
                "comms_wlan": {
                    "type": "string"
                },
                "display_protection": {
                    "type": "string"
                },
                "display_resolution": {
                    "type": "string"
                },
                "display_size": {
                    "type": "string"
                },
                "display_type": {
                    "type": "string"
                },
                "features_sensor": {
                    "type": "string"
                },
                "launch_announce": {
                    "type": "string"
                },
                "launch_status": {
                    "type": "string"
                },
                "maincam_dual": {
                    "type": "string"
                },
                "maincam_features": {
                    "type": "string"
                },
                "maincam_video": {
                    "type": "string"
                },
                "memory_cardslot": {
                    "type": "string"
                },
                "memory_internal": {
                    "type": "string"
                },
                "misc_color": {
                    "type": "string"
                },
                "misc_price": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "network_technology": {
                    "type": "string"
                },
                "picture_cover": {
                    "type": "string"
                },
                "platform_cpu": {
                    "type": "string"
                },
                "platform_gpu": {
                    "type": "string"
                },
                "platform_os": {
                    "type": "string"
                },
                "selfcam_features": {
                    "type": "string"
                },
                "selfcam_single": {
                    "type": "string"
                },
                "selfcam_video": {
                    "type": "string"
                },
                "sound_jack": {
                    "type": "string"
                },
                "sound_loudspeaker": {
                    "type": "string"
                }
            }
        },
...
        "models.Brand": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Opinion": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "phoneID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Phone": {
            "type": "object",
            "properties": {
                "battery_charging": {
                    "type": "string"
                },
                "battery_type": {
                    "type": "string"
                },
                "body_dimensions": {
                    "type": "string"
                },
                "body_sim": {
                    "type": "string"
                },
                "body_weight": {
                    "type": "string"
                },
                "brand": {
                    "$ref": "#/definitions/models.Brand"
                },
                "brandID": {
                    "type": "integer"
                },
                "comms_bluetooth": {
                    "type": "string"
                },
                "comms_gps": {
                    "type": "string"
                },
                "comms_infrared": {
                    "type": "string"
                },
                "comms_nfc": {
                    "type": "string"
                },
                "comms_radio": {
                    "type": "string"
                },
                "comms_usb": {
                    "type": "string"
                },
                "comms_wlan": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "display_protection": {
                    "type": "string"
                },
                "display_resolution": {
                    "type": "string"
                },
                "display_size": {
                    "type": "string"
                },
                "display_type": {
                    "type": "string"
                },
                "features_sensor": {
                    "type": "string"
                },
                "hits": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "launch_announce": {
                    "type": "string"
                },
                "launch_status": {
                    "type": "string"
                },
                "maincam_dual": {
                    "type": "string"
                },
                "maincam_features": {
                    "type": "string"
                },
                "maincam_video": {
                    "type": "string"
                },
                "memory_cardslot": {
                    "type": "string"
                },
                "memory_internal": {
                    "type": "string"
                },
                "misc_color": {
                    "type": "string"
                },
                "misc_price": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "network_technology": {
                    "type": "string"
                },
                "opinions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Opinion"
                    }
                },
                "picture_cover": {
                    "type": "string"
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Picture"
                    }
                },
                "platform_cpu": {
                    "type": "string"
                },
                "platform_gpu": {
                    "type": "string"
                },
                "platform_os": {
                    "type": "string"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Review"
                    }
                },
                "selfcam_features": {
                    "type": "string"
                },
                "selfcam_single": {
                    "type": "string"
                },
                "selfcam_video": {
                    "type": "string"
                },
                "sound_jack": {
                    "type": "string"
                },
                "sound_loudspeaker": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Picture": {
            "type": "object",
            "properties": {
                "UrlPicture": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phoneID": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Review": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phoneID": {
                    "type": "integer"
                },
                "picture": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    }
}